LayoutTests/discovery/video-player-sel-audio-track-expected.txt

Tests HTTPMediaElement Player.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Audio Track added. enabled: true
PASS vid.audioTracks.length>0 is true
PASS vid.audioTracks.length>0 is true
PASS vid.audioTracks.item(0).enabled is true
PASS vid.audioTracks.item(0).enabled is false
PASS successfullyParsed is true

TEST COMPLETE
